requirements.txt
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
xgboost==2.0.0
optuna==3.3.0
feature-engine==1.6.1
shap==0.42.1
requests==2.31.0
beautifulsoup4==4.12.2
yfinance==0.2.36
fredapi==0.5.1
python-dotenv==1.0.0 